CMake: Link against imported targets
As of CMake 3.6, the function pkg_check_modules() supports the argument
IMPORTED_TARGET to provide imported targets which can be passed to
target_link_libraries(). [1]

Although this change might seem to increase the minimum required version
of CMake from 3.4 to 3.6, it actually decreases it from 3.13 to 3.6.
That is because the function target_link_directories() which was used
before actually requires CMake 3.13 or newer in order to work. [2]

[1] https://cmake.org/cmake/help/v3.28/module/FindPkgConfig.html
[2] https://cmake.org/cmake/help/v3.28/command/target_link_directories.html

Signed-off-by: Dennis Menschel <menschel-d@posteo.de>